requirements.txt
attrs
awesome-slugify
click
factory-boy
funcy
ipdb
ipython
jinja2
lxml
networkx
path.py
pycorpora
pyyaml
